feat(solidarity-count): split functions into small files
