+++ packages/webhooks-solidarity-count/src/interfaces/Organizations.ts
@@ -1,0 +1,5 @@
+export enum ORGANIZATIONS {
+  ADVOGADA,
+  MSR,
+  PSICOLOGA
+}
